.requirements.txt
django>=1.11,<1.12
djangorestframework>=3.6,<3.7
djangorestframework-jwt==1.10.0
django-auth-ldap==1.2.12
mysqlclient==1.3.10
django-filter==1.0.4
